add security to notes

+++ routes/notes.py
@@ -1,4 +1,6 @@
 from flask import Blueprint, request, jsonify
+from flask_jwt_extended import jwt_required, get_jwt_identity
+
 from database import db
 from models.Note import Note
 
@@ -7,23 +9,28 @@ notes_bp = Blueprint('notes', __name__)
 
 # Create a note
 @notes_bp.route('/notes', methods=['POST'])
+@jwt_required()
 def create_note():
     data = request.get_json()
     new_note = Note(title=data['title'], content=data['content'])
+    new_note.user_id = get_jwt_identity()
     db.session.add(new_note)
     db.session.commit()
     return jsonify({"message": "Note created!", "note": new_note.to_dict()}), 201
 
 # Get all notes
 @notes_bp.route('/notes', methods=['GET'])
+@jwt_required()
 def get_notes():
-    notes = Note.query.all()
+    user_id = get_jwt_identity()
+    notes = Note.query.filter_by(user_id=user_id).all()
     return jsonify([note.to_dict() for note in notes])
 
 # Get a single note by ID
 @notes_bp.route('/notes/<int:id>', methods=['GET'])
+@jwt_required()
 def get_note(id):
-    note = Note.query.get(id)
+    note = Note.query.filter_by(id=id, user_id=get_jwt_identity()).first()
     if not note:
         return (jsonify
         ({"message": "Note not found!", "status_code": 404}))
@@ -31,8 +38,9 @@ def get_note(id):
 
 # Update a note
 @notes_bp.route('/notes/<int:id>', methods=['PUT'])
+@jwt_required()
 def update_note(id):
-    note = Note.query.get(id)
+    note = Note.query.filter_by(id=id, user_id=get_jwt_identity())
     if not note:
         return (jsonify
     ({"message": "Note not found!", "status_code": 404}))
@@ -47,8 +55,14 @@ def update_note(id):
 
 # Delete a note
 @notes_bp.route('/notes/<int:id>', methods=['DELETE'])
+@jwt_required()
 def delete_note(id):
-    note = Note.query.get_or_404(id)
+    print(f"Delete note: {get_jwt_identity()}")
+    note = Note.query.filter_by(id=id, user_id=get_jwt_identity()).first()
+    print(note, "[[[[[[[]]]]]]]]]]]]]]]]]]]]]]]")
+    if not note:
+        print(note, "[[[[[[[]]]]]]]]]]]]]]]]]]]]]]]")
+        return jsonify({"message": "Note not found!", "status_code": 404})
     db.session.delete(note)
     db.session.commit()
     return jsonify({"message": "Note deleted!"})
